requirements.txt
# Environment management
python-dotenv>=1.0.0,<2.0.0

# LiveKit AI Agents framework (1.3.x)
livekit-agents>=1.3.10
livekit-plugins-aws>=1.3.10
livekit-plugins-silero>=1.3.10

# AWS Transcribe streaming (required for STT)
aws_sdk_transcribe_streaming>=0.1.0

# HTTP client for API calls
httpx>=0.27.0

# Redis client for caching
redis>=5.0.0,<6.0.0

# Note: Pin exact versions in production for reproducible builds
# To generate exact versions: pip freeze > requirements.txt
